Move Integer#positive? and Integer#negative? query methods to Numeric class
By this way Integer, Rational, Float, Fixnum, Bignum classes have the same behaviour
