+++ activesupport/lib/active_support/core_ext/integer.rb
@@ -1,4 +1,3 @@
 require 'active_support/core_ext/integer/multiple'
 require 'active_support/core_ext/integer/inflections'
-require 'active_support/core_ext/integer/inquiry'
 require 'active_support/core_ext/integer/time'
