+++ activesupport/lib/active_support/core_ext/numeric.rb
@@ -1,3 +1,4 @@
 require 'active_support/core_ext/numeric/bytes'
 require 'active_support/core_ext/numeric/time'
+require 'active_support/core_ext/numeric/inquiry'
 require 'active_support/core_ext/numeric/conversions'
